requirements.txt
Flask==2.1.2
google==3.0.0
google-cloud==0.34.0
google-cloud-vision==2.7.2
matplotlib==3.3.4
Pillow==8.2.0
python-dateutil==2.8.1
python-dotenv==0.20.0
regex==2022.3.15
requests==2.25.1
Werkzeug==2.0.2